Crud in admin controller

+++ app/Http/Controllers/Admin/AccountController.php
@@ -11,6 +11,7 @@ use App\Http\Requests\AccountRequest;
 use App\Http\Requests\UpdateAccountRequest;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Redirect;
 
 class AccountController extends Controller
 {
@@ -38,11 +39,11 @@ class AccountController extends Controller
             ->editColumn('is_lock', function ($data) {
                 if ($data->is_lock == 0) 
                 {
-                    return' <a class="fa-thumb-styling fa fa-thumbs-down btn-lg" href="' . route("admin.account.unban_account", $data->id).'"></a>';
+                    return' <a class="fas fa-times btn-lg" href="' . route("admin.account.unban_account", $data->id).'"></a>';
                 }
             else
                 {
-                    return' <a class="fa-thumb-styling fa fa-thumbs-up btn-lg" href="' . route("admin.account.ban_account", $data->id).'"></a>';
+                    return' <a class="fas fa-check-circle btn-lg" href="' . route("admin.account.ban_account", $data->id).'"></a>';
                 }
             })
             ->editColumn('action', function ($data) {
@@ -51,7 +52,7 @@ class AccountController extends Controller
                 <form method="GET" action="' . route('admin.account.delete', $data->id) . '" accept-charset="UTF-8" style="display:inline-block">
                 ' . method_field('GET') .
                     '' . csrf_field() .
-                    '<button type="submit" class="btn btn-danger btn-sm rounded-pill" onclick="return confirm(\'Do you want to delete this kind of room ?\')"><i class="fa-solid fa-trash" title="Delete Mission"></i>Delete</button>
+                    '<button type="submit" class="btn btn-danger btn-sm rounded-pill" onclick="return confirm(\'Do you want to delete this account ?\')"><i class="fa-solid fa-trash" title="Delete Account"></i>Delete</button>
                 </form>
                 ';
             })
@@ -84,8 +85,13 @@ class AccountController extends Controller
     public function delete($id)
     {
         $data = User::find($id);
-        $data->delete();
-        return redirect()->back()->with('message', 'User deleted!');
+        if($data->role->role_name = "admin"){
+            return redirect()->back()->with('message', 'User can not delete!');
+        }
+        else{
+            $data->delete();
+            return redirect()->back()->with('message', 'User deleted!');
+        }
     }
     public function edit($id)
     {
@@ -105,6 +111,14 @@ class AccountController extends Controller
         ]);
         $user->save();
         return redirect('/admin/account/index')->with('success','Update account successfully');
+    }
+    public function unlock_account($id){
+        User::where('id',$id)->update(['is_lock'=>1]);
+        return Redirect::to('/admin/account/index')->with('success','Unlock account successfully');
+    }
+    public function lock_account($id){
+        User::where('id',$id)->update(['is_lock'=>0]);
+        return Redirect::to('/admin/account/index')->with('message','Lock account successfully');
     }
 
 }
